requirements.txt
streamlit
easyocr
pillow
qrcode